\name{GMCB-package}
\alias{GMCB-package}
\alias{GMCB}
\docType{package}
\title{
\packageTitle{GMCB}
}
\description{
\packageDescription{GMCB}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GMCB}
\packageIndices{GMCB}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GMCB}

Maintainer: \packageMaintainer{GMCB}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{

}
\examples{
~~ simple examples of the most important functions ~~
}
