\name{init_naive}
\alias{init_naive}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simple initialization of the GMCB model
}
\description{
This function assumes a multivariate linear regression model for the response.
It initializes parameters with interpretations as regression 
coefficients by zero, parameters with interpretations as regression error 
variances by the sample variances of the response, and the remaining parameters
from user-specified priors.
}
\usage{
init_naive(y, x, lambda.prior, tau.prior, alpha.prior)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{y}{
response matrix with dimensions \emph{n} by \emph{q}. Must have \eqn{q \ge 2}. 
}
  \item{x}{
covariate matrix with dimensions \emph{n} by \emph{p}.
}
  \item{lambda.prior}{
either a vector of length 4 or a matrix with dimensions 4 by \emph{pq}. 
}
  \item{tau.prior}{
either a vector of length 4 or a matrix with dimensions 4 by \emph{q(q-1)/2}.
}
  \item{alpha.prior}{
vector of length 2 giving the minimum and maximum values for the uniform prior
on the penalty exponents.
}
}
\details{
The initializing values for the regularization parameters and the penalty 
exponents are generated by the function \code{nuisance_init}. See the documentation
for \code{nuisance_init} for details on the specification of \code{lambda.prior}
and \code{tau.prior}.
}
\value{
A list including the following componets:
  \item{b.init}{a vector of length \emph{pq}, the initializing values
                for the mean regression coefficient matrix.}
  \item{delta.init}{a vector of length \emph{q(q-1)/2}, the initializing values
                    for the covariance parameters with interpretations as 
                    regression coefficients.}
  \item{lambda.init}{a vector of length \emph{pq}, the initializing values for the 
                     regularization parameters for the mean regression coefficient matrix.}
  \item{tau.init}{a vector of length \emph{q(q-1)/2}, the initializing values for the 
                  regularization parameters for the covariance parameters.}
  \item{alpha.b.init}{scalar, the initializing value for the penalty exponent
                      for the mean.}
  \item{alpha.d.init}{scalar, the initializing value for the penalty exponent
                      for the covariance parameters.}
  \item{gamma.init}{a vector of length \emph{q}, the initializing values for 
                    the covariance parameters with interpretations as 
                    regression error variances.}
  \item{gamma.prior}{a vector of length two, giving the shape and rate 
                     parameters for the inverse Gamma prior on the 
                     covariance parameters with interpretations as 
                     regression error variances.}
}
%\references{
%% ~put references to the literature/web site here ~
%}
%\author{
%Christina Zhao
%}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{init_freqbridge_nocovariates}}, \code{\link{nuisance.init}}, 
\code{\link{ols.init}}
}
\examples{

library(mvtnorm)

n <- 100
p <- 5
q <- 5

rho <- 0.7

x.mats <- matgen(p, rho, type = "AR")
y.mats <- matgen(q, rho, type = "AR")

x <- rmvnorm(n, mean = rep(0, p), sigma = x.mats$cov.mat)
b <- matrix(rnorm(p*q), nrow = p)
y <- x \%*\% b + rmvnorm(n, mean = rep(0, q), sigma = y.mats$cov.mat)

lambda.prior <- c(1, 1, 40, 0.5)
tau.prior <- c(1, 1, 40, 0.5)
alpha.prior <- c(0.5, 4)

initialize <- init_naive(y, x, lambda.prior, tau.prior, alpha.prior)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
